+++ tee-worker/identity/client-sdk/packages/client-sdk/src/lib/requests/index.ts
@@ -3,6 +3,7 @@ export { linkIdentityCallback } from './link-identity-callback.request';
 export { setIdentityNetworks } from './set-identity-networks.request';
 export { requestBatchVC } from './request-batch-vc.request';
 export { remark } from './remark.request';
+export { createAccountStore } from './create-account-store.request';
 export { getIdGraph } from './get-id-graph.request';
 export { getIdGraphHash } from './get-id-graph-hash';
 export { getLastRegisteredEnclave } from './get-last-registered-enclave';